Use stdlib router
This is an important step to introducing multipart uploads, which make
a complete hash of the nice clean routing. It really feels like S3
started out with a clean restish API then they just started stapling
stuff to the side one piece at a time until they had a really big mess.

This also includes a potentially contentious change to unexport the
GoFakeS3 handler functions. I've done this in order to limit the surface
of supported API for now, it's not something I think should necessarily be
permanent, it's more just to paint a smaller compatibility target that
we have to support, at the risk of breaking existing code that might be
calling these handlers directly (though the suggested upgrade path would
be to call the backend directly anyway).

Things might be in slightly too much flux at the moment (and I think the
README already does a good job of making the level of stability a user
could expect clear), but we should definitely look at version numbering
at some point in the future and a stable API surface is a critical part
of that.