Initial backend multipart implementation for bolt.
Currently not finished (list features are unimplemented) and in-memory, but it
generally works.
This is more suitable as an example on how an implementor can make use
of the new API to pull control into the backend.

This disregards reuse of content chunks etc. (for now?) With little
thought to an implementor they can decide how to persist data if required.

It also has a dirty hack to get the ETag, as `db.PutObject` just returns the
`VersionID`. Maybe this is a future optimization for the API, but I could
have gone to the source anyways, because I AM in the backend here. Just took
a (very inefficient, but boring) shortcut.

It might be a good idea to create a `ProxyBackend` of some sorts, that can
creeate this in-memory API implementation for other, 'standard' backends.
That way we could get rid of most of the switching code in `gofakes3.go`.

Some debug printfs still inside.

Upload and download was again tested via aws cli.